+++ crates/deps/reader/src/async_kind.rs
@@ -1,0 +1,8 @@
+#[derive(PartialEq)]
+pub enum AsyncKind {
+    None,
+    Action,
+    ActionWithProgress,
+    Operation,
+    OperationWithProgress,
+}
