+++ crates/deps/reader/src/lib.rs
@@ -18,6 +18,7 @@ mod type_name;
 mod type_reader;
 mod type_tree;
 mod workspace;
+mod async_kind;
 
 pub use blob::*;
 pub use codes::*;
@@ -39,3 +40,4 @@ pub use type_name::*;
 pub use type_reader::*;
 pub use type_tree::*;
 pub use workspace::*;
+pub use async_kind::*;
